pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: Apache Software License",
]
name = "nomad-uibk-plugin"
description = "UIBK schema and parser collection for NOMAD plattform."
version = "0.1.0"
readme = "README.rst"
requires-python = ">=3.9"
authors = [
    { name = "Fabian Schöppach", email = "fabian.schoeppach@physik.hu-berlin.de" },
]
maintainers = [
    { name = "Fabian Schöppach", email = "fabian.schoeppach@physik.hu-berlin.de" },
]
license = { file = "LICENSE" }
dependencies = [
    "nomad-lab>=1.2.2dev578",
    "nomad-measurements>=0.0.05",
    "nomad-material-processing>=0.0.10",
    "ifm-image-defect-detection @ git+https://github.com/csav1974/IFM-Image-Defect-Detection-and-Classification.git@pyproject-ready"
    ]

[project.urls]
Repository = "https://github.com/fabianschoeppach/nomad-uibk-plugin"

[tool.uv]
index-url = "https://gitlab.mpcdf.mpg.de/api/v4/projects/2187/packages/pypi/simple"

[project.optional-dependencies]
dev = ["ruff", "pytest", "structlog"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4


[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # isort
    "I",
    # pylint
    "PL",
]

ignore = [
    "F403", # 'from module import *' used; unable to detect undefined names
]

fixable = ["ALL"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# this is entirely optional, you can remove this if you wish to
[tool.ruff.format]
# use single quotes for strings.
quote-style = "single"

# indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[project.entry-points.'nomad.plugin']
# schema
sample = "nomad_uibk_plugin.schema_packages:sample"
xrfschema = "nomad_uibk_plugin.schema_packages:xrfschema"
ifmschema = "nomad_uibk_plugin.schema_packages:ifmschema"
# parser
xrfparser = "nomad_uibk_plugin.parsers:xrfparser"
# unsure
#microcellschema = "nomad_uibk_plugin.schema_packages:microcellschema"
#ifmparser = "nomad_uibk_plugin.parsers:ifmparser"
#ebicparser = "nomad_uibk_plugin.parsers:ebicparser"
